refactor: remove service interfaces and fix game character validation when creating combo from playlist
