+++ backend/src/main/java/com/fgc/combo/companion/repository/GameCharacterRepository.java
@@ -11,5 +11,8 @@ import com.fgc.combo.companion.model.GameCharacter;
 public interface GameCharacterRepository extends JpaRepository<GameCharacter, Long> {
     @Query(value = "SELECT * FROM game_characters WHERE game = CAST(:game as gametypes)", nativeQuery = true)
     Set<GameCharacter> findAllByGame(@Param("game") String game);
+    
+    @Query(value = "SELECT count(*) > 0 FROM game_characters WHERE game = CAST(:game as gametypes) AND code = :code", nativeQuery = true)
+    boolean existsByCodeAndGame(String code, String game);
 
 }
