+++ backend/src/main/java/com/fgc/combo/companion/service/CustomUserDetailsServiceImpl.java
@@ -1,4 +1,4 @@
-package com.fgc.combo.companion.service.impl;
+package com.fgc.combo.companion.service;
 
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
