+++ backend/src/main/java/com/fgc/combo/companion/service/GameCharactersService.java
@@ -2,9 +2,33 @@ package com.fgc.combo.companion.service;
 
 import java.util.Set;
 
-import com.fgc.combo.companion.enums.ComboGameTypes;
-import com.fgc.combo.companion.model.GameCharacter;
+import org.springframework.stereotype.Service;
 
-public interface GameCharactersService {
-  Set<GameCharacter> listGameCharactersByGame(ComboGameTypes game);
+import com.fgc.combo.companion.enums.ComboGameTypes;
+import com.fgc.combo.companion.exception.BadRequestException;
+import com.fgc.combo.companion.model.GameCharacter;
+import com.fgc.combo.companion.repository.GameCharacterRepository;
+
+@Service
+public class GameCharactersService {
+
+  private final GameCharacterRepository gameCharacterRepository;
+
+  public GameCharactersService(
+      GameCharacterRepository gameCharacterRepository) {
+    this.gameCharacterRepository = gameCharacterRepository;
+  }
+
+  public Set<GameCharacter> listGameCharactersByGame(ComboGameTypes comboGameType) {
+    return this.gameCharacterRepository.findAllByGame(comboGameType.name());
+  }
+
+  public boolean isGameCharacterValid(String comboCharacter, String comboGameType) {
+    if (!this.gameCharacterRepository.existsByCodeAndGame(comboCharacter, comboGameType)) {
+      throw new BadRequestException("Game Character does not exists!");
+    }
+
+    return true;
+  }
+
 }
